# os/bluestore: _dump_onode() don't prolongate Onode anymore.