@liewegas 2018-01-09
http://pulpito.ceph.com/sage-2018-01-08_20:16:48-rados-wip-sage3-testing-2018-01-08-1100-distro-basic-smithi/